% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{KMPSearch}
\alias{KMPSearch}
\title{KMPSearch}
\usage{
KMPSearch(Query, Target, Next)
}
\arguments{
\item{Query}{query string array}

\item{Target}{target string array}

\item{Next}{next array generated by buildNext}
}
\value{
KMPSearch
}
\description{
KMPSearch
}
