% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{match}
\alias{match}
\title{match}
\usage{
match(s, p, i)
}
\arguments{
\item{s}{first number}

\item{p}{second number}

\item{i}{number}
}
\value{
match index
}
\description{
string match.
}
\examples{
require(Bruteforce)
s <- ("abc")
p <- ("b")
i <- 1
print(match("abc", "b", 1))

}
