% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{rabin_karp}
\alias{rabin_karp}
\title{rabin_karp}
\usage{
rabin_karp(s, p)
}
\arguments{
\item{s}{first number}

\item{p}{second number}
}
\value{
match index
}
\description{
rabin_karp
}
\examples{
require(Bruteforce)
s <- ("abc")
p <- ("b")
print(rabin_karp("abc", "b"))

}
